Added map version picker
Moved maps to resorces
